+++ lib/lightning/accounts/user.ex
@@ -34,6 +34,10 @@ defmodule Lightning.Accounts.User do
     field :scheduled_deletion, :utc_datetime
     field :github_oauth_token, Lightning.Encrypted.Map, redact: true
 
+    field :emails_preference, Ecto.Enum,
+      values: [:critical, :any],
+      default: :critical
+
     has_one :user_totp, Lightning.Accounts.UserTOTP
     has_many :credentials, Lightning.Credentials.Credential
     has_many :oauth_clients, Lightning.Credentials.OauthClient
@@ -52,7 +56,8 @@ defmodule Lightning.Accounts.User do
       :first_name,
       :last_name,
       :email,
-      :password
+      :password,
+      :emails_preference
     ])
     |> validate_name()
     |> validate_email()
@@ -66,7 +71,9 @@ defmodule Lightning.Accounts.User do
     password: :string,
     hashed_password: :string,
     disabled: :boolean,
-    scheduled_deletion: :utc_datetime
+    scheduled_deletion: :utc_datetime,
+    emails_preference:
+      Ecto.ParameterizedType.init(Ecto.Enum, values: [:critical, :any])
   }
 
   @doc """