+++ priv/repo/migrations/20240530115147_add_email_preference_to_users.exs
@@ -1,0 +1,9 @@
+defmodule Lightning.Repo.Migrations.AddEmailsPreferenceToUsers do
+  use Ecto.Migration
+
+  def change do
+    alter table(:users) do
+      add :emails_preference, :string, default: "critical"
+    end
+  end
+end